Rename: 패키지를 적절히 이동한다.
