+++ src/main/java/org/prgms/springbootbasic/repository/customer/CustomerFileRepository.java
@@ -1,4 +1,4 @@
-package org.prgms.springbootbasic.repository;
+package org.prgms.springbootbasic.repository.customer;
 
 import org.prgms.springbootbasic.common.file.CustomerCsvFileManager;
 import org.prgms.springbootbasic.domain.customer.Customer;
