+++ src/main/java/org/prgms/springbootbasic/repository/customer/CustomerMemoryRepository.java
@@ -1,4 +1,4 @@
-package org.prgms.springbootbasic.repository;
+package org.prgms.springbootbasic.repository.customer;
 
 import org.prgms.springbootbasic.domain.customer.Customer;
 import org.springframework.context.annotation.Profile;
