+++ src/main/java/org/prgms/springbootbasic/repository/voucher/VoucherMemoryRepository.java
@@ -1,4 +1,4 @@
-package org.prgms.springbootbasic.repository;
+package org.prgms.springbootbasic.repository.voucher;
 
 import lombok.extern.slf4j.Slf4j;
 import org.prgms.springbootbasic.domain.voucher.VoucherPolicy;
