+++ src/main/java/org/prgms/springbootbasic/service/CustomerService.java
@@ -2,7 +2,7 @@ package org.prgms.springbootbasic.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.prgms.springbootbasic.domain.customer.Customer;
-import org.prgms.springbootbasic.repository.CustomerRepository;
+import org.prgms.springbootbasic.repository.customer.CustomerRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
